Move precompilation to LocalSampler
